Fix bug in the mesh collision model and clean a bit.

+++ include/smExternal/moller2.h
@@ -1,8 +1,8 @@
 /********************************************************/
 /* AABB-triangle overlap test code                      */
 /* by Tomas Akenine-Möller                              */
-/* Function: int triBoxOverlap(float boxcenter[3],      */
-/*          float boxhalfsize[3],float triverts[3][3]); */
+/* Function: int triBoxOverlap(double boxcenter[3],      */
+/*          double boxhalfsize[3],double triverts[3][3]); */
 /* History:                                             */
 /*   2001-03-05: released the code in its first version */
 /*   2001-06-18: changed the order of the tests, faster */
@@ -39,10 +39,10 @@
   if(x2<min) min=x2;\
   if(x2>max) max=x2;
 
-int planeBoxOverlap(float normal[3],float d, float maxbox[3])
+int planeBoxOverlap(double normal[3],double d, double maxbox[3])
 {
   int q;
-  float vmin[3],vmax[3];
+  double vmin[3],vmax[3];
   for(q=X;q<=Z;q++)
   {
     if(normal[q]>0.0f)
@@ -109,7 +109,7 @@ int planeBoxOverlap(float normal[3],float d, float maxbox[3])
 	rad = fa * boxhalfsize[X] + fb * boxhalfsize[Y];   \
 	if(min>rad || max<-rad) return 0;
 
-int triBoxOverlap(float boxcenter[3],float boxhalfsize[3],float triverts[3][3])
+int triBoxOverlap(double boxcenter[3],double boxhalfsize[3],double triverts[3][3])
 {
 
   /*    use separating axis theorem to test overlap between triangle and box */
@@ -119,9 +119,9 @@ int triBoxOverlap(float boxcenter[3],float boxhalfsize[3],float triverts[3][3])
   /*    2) normal of the triangle */
   /*    3) crossproduct(edge from tri, {x,y,z}-directin) */
   /*       this gives 3x3=9 more tests */
-   float v0[3],v1[3],v2[3];
-   float min,max,d,p0,p1,p2,rad,fex,fey,fez;
-   float normal[3],e0[3],e1[3],e2[3];
+   double v0[3],v1[3],v2[3];
+   double min,max,d,p0,p1,p2,rad,fex,fey,fez;
+   double normal[3],e0[3],e1[3],e2[3];
 
    /* This is the fastest branch on Sun */
    /* move everything so that the boxcenter is in (0,0,0) */
